EDMT-365 refactor : 기존 생기부 관리 폴더 record-manage로 폴더 이름 수정
