+++ src/app/(main)/(record)/manage-behavior/page.tsx
@@ -1,4 +1,4 @@
-import BehaviorRecordTable from '@/domains/record/components/student-manage/behavior-record-table';
+import BehaviorRecordTable from '@/domains/record/components/record-manage/behavior-record-table';
 import { createPageMetadata } from '@/shared/constants/metadata';
 
 export const metadata = createPageMetadata('manageStudent', { url: '/manage-behavior' });
