+++ src/app/(main)/(record)/manage-career/page.tsx
@@ -1,4 +1,4 @@
-import CareerRecordTable from '@/domains/record/components/student-manage/career-record-table';
+import CareerRecordTable from '@/domains/record/components/record-manage/career-record-table';
 import { createPageMetadata } from '@/shared/constants/metadata';
 
 export const metadata = createPageMetadata('manageStudent', { url: '/manage-career' });
