+++ src/app/(main)/(record)/manage-student/page.tsx
@@ -1,4 +1,4 @@
-import FreeRecordTable from '@/domains/record/components/student-manage/free-record-table';
+import FreeRecordTable from '@/domains/record/components/record-manage/free-record-table';
 import { createPageMetadata } from '@/shared/constants/metadata';
 
 export const metadata = createPageMetadata('manageStudent', { url: '/manage-free' });
