+++ src/app/(main)/(record)/manage-subject/page.tsx
@@ -1,4 +1,4 @@
-import SubjectRecordTable from '@/domains/record/components/student-manage/subject-record-table';
+import SubjectRecordTable from '@/domains/record/components/record-manage/subject-record-table';
 import { createPageMetadata } from '@/shared/constants/metadata';
 
 export const metadata = createPageMetadata('manageStudent', { url: '/manage-subject' });
