+++ src/domains/record/components/record-manage/behavior-record-table.tsx
