+++ src/domains/record/components/record-manage/career-record-table.tsx
