+++ src/domains/record/components/record-manage/club-record-table.tsx
