+++ src/domains/record/components/record-manage/record-body.tsx
