+++ src/domains/record/components/record-manage/record-detail-add.tsx
