+++ src/domains/record/components/record-manage/record-detail-edit.tsx
