+++ src/domains/record/components/record-manage/record-detail-view.tsx
